% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFDocument.R
\name{add_polyline}
\alias{add_polyline}
\title{Add a polyline}
\arguments{
\item{pdfdoc}{PDF document. An R6 object of class \code{PDFDocument}. Created with \code{create_pdfdoc()} or \code{PDFDocument$new()}}

\item{xs, ys}{vectors of coordinates of each point along the polyline}

\item{stroke}{Colour. May be given as hex-colour \code{'#123456'}, RGB vector \code{c(0, 200, 12)}, or RGBA vector e.g. \code{c(0, 156, 255, 255)}}

\item{linewidth}{line width (pixels)}

\item{linetype}{Integer from 0 (solid)u to 6 (twodash)}
}
\value{
PDF document with line added
}
\description{
Adds a line to the current document
}
\section{R6 Usage}{
 $add_polyline()
}

