% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFDocument.R
\name{add_rect}
\alias{add_rect}
\title{Add rectangle}
\arguments{
\item{pdfdoc}{PDF document. An R6 object of class \code{PDFDocument}. Created with \code{create_pdfdoc()} or \code{PDFDocument$new()}}

\item{x, y, width, height}{location and size of rectangle}

\item{fill, stroke}{Colours. May be given as hex-colour \code{'#123456'}, RGB vector \code{c(0, 200, 12)}, or RGBA vector e.g. \code{c(0, 156, 255, 255)}}

\item{linewidth}{line width (pixels)}

\item{linetype}{Integer from 0 (solid)u to 6 (twodash)}
}
\value{
PDF document with rectangle added
}
\description{
Adds rectangle to the current document
}
\section{R6 Usage}{
 $add_rect()
}

