% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFDocument.R
\name{add_text}
\alias{add_text}
\title{Add text}
\arguments{
\item{pdfdoc}{PDF document. An R6 object of class \code{PDFDocument}. Created with \code{create_pdfdoc()} or \code{PDFDocument$new()}}

\item{text}{text to add (string)}

\item{x, y}{location of text}

\item{size}{font size. Default: 12}

\item{angle}{rotation angle measured in degrees anti-clockwise. Default: 0}

\item{fill, stroke}{colours. May be given as hex-colour \code{'#123456'}, RGB vector \code{c(0, 200, 12)}, or RGBA vector e.g. \code{c(0, 156, 255, 255)}}

\item{text_mode}{Integer. 0 = fill, 1 = stroke, 2 = fill and stroke. Default: 0}

\item{linewidth}{line width (pixels)}

\item{linetype}{Integer from 0 (solid) to 6 (twodash)}
}
\value{
PDF document with line added
}
\description{
Adds text to the current document
}
\section{R6 Usage}{
 $add_text()
}

