% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_named}
\alias{all_named}
\title{Check if all elements are named}
\usage{
all_named(x)
}
\arguments{
\item{x}{object e.g. list, vector}
}
\value{
TRUE if all elements are named, otherwise FALSE
}
\description{
Check if all elements are named
}
