% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf-stream.R
\name{as.character.pdf_stream}
\alias{as.character.pdf_stream}
\title{Convert \code{pdf_stream} to character}
\usage{
\method{as.character}{pdf_stream}(x, ...)
}
\arguments{
\item{x}{\code{pdf_stream} object}

\item{...}{ignored}
}
\value{
character string representation of a pdf stream object
}
\description{
Convert \code{pdf_stream} to character
}
