% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{as.character.pdf_translate}
\alias{as.character.pdf_translate}
\alias{as.character.pdf_rotate}
\alias{as.character.pdf_scale}
\alias{as.character.pdf_transform_list}
\title{Convert transform to character string}
\usage{
\method{as.character}{pdf_translate}(x, ...)

\method{as.character}{pdf_rotate}(x, ...)

\method{as.character}{pdf_scale}(x, ...)

\method{as.character}{pdf_transform_list}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{ignored}
}
\value{
string
}
\description{
Convert transform to character string
}
