% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFDict.R
\name{as.character.PDFDict}
\alias{as.character.PDFDict}
\title{S3 method for converting PDFDict to character}
\usage{
\method{as.character}{PDFDict}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{unused}
}
\description{
S3 method for converting PDFDict to character
}
