% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFDocument.R
\name{as.character.PDFDocument}
\alias{as.character.PDFDocument}
\title{S3 method for converting PDFDict to character}
\usage{
\method{as.character}{PDFDocument}(x, ...)
}
\arguments{
\item{x}{PDFDocument object}

\item{...}{unused}
}
\description{
S3 method for converting PDFDict to character
}
