% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFStream.R
\name{as.character.PDFStream}
\alias{as.character.PDFStream}
\title{S3 method for converting PDFStream to character}
\usage{
\method{as.character}{PDFStream}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{unused}
}
\description{
S3 method for converting PDFStream to character
}
