% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif.R
\name{as.character.stream}
\alias{as.character.stream}
\title{Convert stream to character}
\usage{
\method{as.character}{stream}(x, ...)
}
\arguments{
\item{x}{pif_stream object}

\item{...}{ignored}
}
\value{
character string
}
\description{
Convert stream to character
}
