% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{create_pdf}
\alias{create_pdf}
\title{Create an new PDF}
\usage{
create_pdf(
  width = 400,
  height = 400,
  title = NULL,
  author = NULL,
  creator = "minipdf/R",
  creation_date = strftime(Sys.time(), format = "D:\%Y\%m\%d\%H\%M")
)
}
\arguments{
\item{width, height}{page size in pixels}

\item{title, author, creator, creation_date}{Document-level metainformation
about this file. Set value to NULL to exclude from PDF.}
}
\value{
\code{pdf_doc} object (i.e. a named list)
}
\description{
Create an new PDF
}
\examples{
create_pdf()
}
