% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFDocument.R
\name{create_pdfdoc}
\alias{create_pdfdoc}
\title{Create a new PDF Document}
\usage{
create_pdfdoc(width = 400, height = 400, font = "Helvetica",
  version = 1.2)
}
\arguments{
\item{width, height}{Page dimensions in pixels}

\item{font}{name of font. One of: Times-Roman, Times-Bold, Times-Italic,
Times-BoldItalic, Helvetica, Helvetica-Bold, Helvetica-Oblique, \
Helvetica-BoldOblique, Courier, Courier-Bold, Courier-Oblique,
Courier-BoldOblique, Symbol, ZapfDingbats}

\item{version}{PDF standard version number. default 1.2}
}
\value{
Blank PDF document
}
\description{
Create a new PDF Document
}
\section{R6 Usage}{
 $new()
}

