% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif.R
\name{create_pif}
\alias{create_pif}
\title{Create an empty shell for the PDF intermediate format}
\usage{
create_pif(...)
}
\arguments{
\item{...}{options}
}
\value{
List with attributes. List items are PDF objects.  Attributes
        are PDF settings
}
\description{
Create an empty shell for the PDF intermediate format
}
\examples{
create_pif()

}
