% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFDocument.R
\name{create_rotate_transform_about_coord}
\alias{create_rotate_transform_about_coord}
\title{Create rotation transform about a given coord}
\usage{
create_rotate_transform_about_coord(x, y, angle)
}
\arguments{
\item{x, y}{centre of rotation}

\item{angle}{rotation angle in degrees}
}
\value{
transofmration string
}
\description{
Create rotation transform about a given coord
}
