% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{dec_ascii85}
\alias{dec_ascii85}
\title{Decode a ASCII85-encoded character string into binary data}
\usage{
dec_ascii85(s)
}
\arguments{
\item{s}{string}
}
\value{
raw vector
}
\description{
Decode a ASCII85-encoded character string into binary data
}
