% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFDict.R
\name{dict}
\alias{dict}
\title{Wrapper to help create a PDFDict object}
\usage{
dict(...)
}
\arguments{
\item{...}{arguments passed to \code{PDFDict$new()}}
}
\value{
PDFDict R6 object
}
\description{
Wrapper to help create a PDFDict object
}
