% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{enc_ascii85}
\alias{enc_ascii85}
\title{Encode binary data into ASCII85-encoded string}
\usage{
enc_ascii85(rv)
}
\arguments{
\item{rv}{raw,integer or numeric vector with values in range [0, 255]}
}
\value{
ASCII85-encoded string
}
\description{
Encode binary data into ASCII85-encoded string
}
