% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{enc_rle}
\alias{enc_rle}
\title{Runlength encode}
\usage{
enc_rle(rv)
}
\arguments{
\item{rv}{raw, integer or numeric vector containing byte values in the
range [0, 255]}
}
\value{
raw vector of run-length encoded valeus as per PDF docs
        Sect 7.4.5. "RunLengthDecode filter"
}
\description{
Encode raw byte data with Run Length Encoding
[controlbyte, sequence, controlbyte, sequence, end-of-data]
Control byte:
  0-127 indicates the following (n + 1) bytes should be copied verbatim
  129-255 The following 257-n bytes are all repeats of the next character
  128 is the end-of-data marker
}
