% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFDocument.R
\name{get_path_paint_style}
\alias{get_path_paint_style}
\title{Get the path painting style based upon which of fill/stroke are NULL}
\usage{
get_path_paint_style(fill, stroke)
}
\arguments{
\item{fill, stroke}{colours}
}
\value{
one of 'n', 's', 'f', 'b'
}
\description{
Get the path painting style based upon which of fill/stroke are NULL
}
