% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{image_to_bytes}
\alias{image_to_bytes}
\title{Convert image to bytes for RGB + Alpha (or Gray + Alpha)}
\usage{
image_to_bytes(im)
}
\arguments{
\item{im}{image. numeric values in range [0, 255]
\itemize{
  \item{Grayscale - numeric matrix}
  \item{Grayscale with alpha - numeric array with 2 planes}
  \item{RGB - numeric array with 3 planes}
  \item{RGB with alpha - numeric array with 4 planes}
  \item{RGB with alpha - nativeraster}
}}
}
\value{
named list of 3 elements: type, rgb (or g) and alpha
}
\description{
Convert image to bytes for RGB + Alpha (or Gray + Alpha)
}
