% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minipdf-package.R
\docType{package}
\name{minipdf-package}
\alias{minipdf}
\alias{minipdf-package}
\title{minipdf: PDF Document Creator}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

PDF is an ISO standard file format for formatting text and images in documents. At its core, these documents are sequences of objects defined in plain text. This package allows for the creation of PDF documents at a very low level without any library or graphics device depenedencies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coolbutuseless/minipdf}
}

}
\author{
\strong{Maintainer}: Mike Cheng \email{mikefc@coolbutuseless.com} [copyright holder]

}
\keyword{internal}
