% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{pdf_add}
\alias{pdf_add}
\title{Add a 'dict' or stream to the 'pdf'}
\usage{
pdf_add(doc, x, pos = NULL)
}
\arguments{
\item{doc}{pdf_doc}

\item{x}{pdf_dict or pdf_stream}

\item{pos}{position at which to add item. Item currently at this
position will be moved to next position}
}
\value{
pdf
}
\description{
Add a 'dict' or stream to the 'pdf'
}
