% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{pdf_append}
\alias{pdf_append}
\title{Append a 'dict' to the 'pdf'}
\usage{
pdf_append(doc, dict)
}
\arguments{
\item{doc}{pdf_doc}

\item{dict}{pdf_dict}
}
\value{
pdf
}
\description{
Append a 'dict' to the 'pdf'
}
