% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFDocument.R
\name{pdf_doc}
\alias{pdf_doc}
\title{Wrapper to help create a PDFDocument object}
\usage{
pdf_doc(..., width = 400, height = 400, fontname = "Helvetica",
  version = 1.2)
}
\arguments{
\item{...}{PDF objects}

\item{width, height}{document dimensions. Default 400x400}

\item{fontname}{default: Helvetica. Choose one of the standard PDF
fonts: Times-Roman, Times-Bold,
Times-Italic, Times-BoldItalic, Helvetica, Helvetica-Bold, Helvetica-Oblique,
Helvetica-BoldOblique, Courier, Courier-Bold, Courier-Oblique, Courier-BoldOblique,
Symbol, ZapfDingbats}

\item{version}{Default: 1.2.  Doesn't do anything other than set the string at the
beginning of the PDF document.}
}
\value{
PDFDocument R6 object
}
\description{
Wrapper to help create a PDFDocument object
}
