% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{pdf_render}
\alias{pdf_render}
\title{export pdf}
\usage{
pdf_render(doc, filename = NULL)
}
\arguments{
\item{doc}{pdf_doc}

\item{filename}{defualt: NULL -> return string}
}
\value{
string
}
\description{
export pdf
}
