% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFStreamX.R
\docType{data}
\name{PDFClipRect}
\alias{PDFClipRect}
\title{PDF ClipRect stream object creator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PDFClipRect
}
\description{
This is the R6 class representing the ClipRect stream object.
}
\details{
\itemize{
\item{x,y coordinates}
\item{width,height rectangle width and height}
\item{clip_rect clipping rectangle for this object.  Must be a vector of4 numbers giving x, y, width and height. default: NULL (no clipping).}
\item{... extra arguments ignored}
}
}
\keyword{datasets}
