% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFStreamX.R
\name{PDFCustom}
\alias{PDFCustom}
\title{PDF Custom stream object creator}
\description{
PDF Custom stream object creator

PDF Custom stream object creator
}
\details{
This is the R6 class representing the custom stream object.
}
\section{Super class}{
\code{\link[minipdf:PDFStream]{minipdf::PDFStream}} -> \code{PDFCustom}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PDFCustom-new}{\code{PDFCustom$new()}}
\item \href{#method-PDFCustom-update}{\code{PDFCustom$update()}}
\item \href{#method-PDFCustom-get_geom_spec}{\code{PDFCustom$get_geom_spec()}}
\item \href{#method-PDFCustom-clone}{\code{PDFCustom$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="minipdf" data-topic="PDFStream" data-id="as_character"><a href='../../minipdf/html/PDFStream.html#method-PDFStream-as_character'><code>minipdf::PDFStream$as_character()</code></a></span></li>
<li><span class="pkg-link" data-pkg="minipdf" data-topic="PDFStream" data-id="as_object"><a href='../../minipdf/html/PDFStream.html#method-PDFStream-as_object'><code>minipdf::PDFStream$as_object()</code></a></span></li>
<li><span class="pkg-link" data-pkg="minipdf" data-topic="PDFStream" data-id="clip_polygon"><a href='../../minipdf/html/PDFStream.html#method-PDFStream-clip_polygon'><code>minipdf::PDFStream$clip_polygon()</code></a></span></li>
<li><span class="pkg-link" data-pkg="minipdf" data-topic="PDFStream" data-id="clip_rect"><a href='../../minipdf/html/PDFStream.html#method-PDFStream-clip_rect'><code>minipdf::PDFStream$clip_rect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="minipdf" data-topic="PDFStream" data-id="copy"><a href='../../minipdf/html/PDFStream.html#method-PDFStream-copy'><code>minipdf::PDFStream$copy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="minipdf" data-topic="PDFStream" data-id="fill"><a href='../../minipdf/html/PDFStream.html#method-PDFStream-fill'><code>minipdf::PDFStream$fill()</code></a></span></li>
<li><span class="pkg-link" data-pkg="minipdf" data-topic="PDFStream" data-id="fontsize"><a href='../../minipdf/html/PDFStream.html#method-PDFStream-fontsize'><code>minipdf::PDFStream$fontsize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="minipdf" data-topic="PDFStream" data-id="get_attrib"><a href='../../minipdf/html/PDFStream.html#method-PDFStream-get_attrib'><code>minipdf::PDFStream$get_attrib()</code></a></span></li>
<li><span class="pkg-link" data-pkg="minipdf" data-topic="PDFStream" data-id="get_transform_spec"><a href='../../minipdf/html/PDFStream.html#method-PDFStream-get_transform_spec'><code>minipdf::PDFStream$get_transform_spec()</code></a></span></li>
<li><span class="pkg-link" data-pkg="minipdf" data-topic="PDFStream" data-id="linetype"><a href='../../minipdf/html/PDFStream.html#method-PDFStream-linetype'><code>minipdf::PDFStream$linetype()</code></a></span></li>
<li><span class="pkg-link" data-pkg="minipdf" data-topic="PDFStream" data-id="linewidth"><a href='../../minipdf/html/PDFStream.html#method-PDFStream-linewidth'><code>minipdf::PDFStream$linewidth()</code></a></span></li>
<li><span class="pkg-link" data-pkg="minipdf" data-topic="PDFStream" data-id="print"><a href='../../minipdf/html/PDFStream.html#method-PDFStream-print'><code>minipdf::PDFStream$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="minipdf" data-topic="PDFStream" data-id="rotate"><a href='../../minipdf/html/PDFStream.html#method-PDFStream-rotate'><code>minipdf::PDFStream$rotate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="minipdf" data-topic="PDFStream" data-id="scale"><a href='../../minipdf/html/PDFStream.html#method-PDFStream-scale'><code>minipdf::PDFStream$scale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="minipdf" data-topic="PDFStream" data-id="stroke"><a href='../../minipdf/html/PDFStream.html#method-PDFStream-stroke'><code>minipdf::PDFStream$stroke()</code></a></span></li>
<li><span class="pkg-link" data-pkg="minipdf" data-topic="PDFStream" data-id="text_mode"><a href='../../minipdf/html/PDFStream.html#method-PDFStream-text_mode'><code>minipdf::PDFStream$text_mode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="minipdf" data-topic="PDFStream" data-id="translate"><a href='../../minipdf/html/PDFStream.html#method-PDFStream-translate'><code>minipdf::PDFStream$translate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFCustom-new"></a>}}
\if{latex}{\out{\hypertarget{method-PDFCustom-new}{}}}
\subsection{Method \code{new()}}{
Initialize a stream object representing a custom object with text
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFCustom$new(text, new_graphics_state = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{text string}

\item{\code{new_graphics_state}}{Should the object be drawn in its own local
graphics state? default: TRUE}

\item{\code{...}}{initial named attributes of this object
Update the attributes of this object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFCustom-update"></a>}}
\if{latex}{\out{\hypertarget{method-PDFCustom-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFCustom$update(text, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{text}

\item{\code{...}}{other atttributes}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFCustom-get_geom_spec"></a>}}
\if{latex}{\out{\hypertarget{method-PDFCustom-get_geom_spec}{}}}
\subsection{Method \code{get_geom_spec()}}{
Fetch the character representation of this geometry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFCustom$get_geom_spec()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFCustom-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PDFCustom-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFCustom$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
