% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFDict.R
\docType{data}
\name{PDFDict}
\alias{PDFDict}
\title{PDF Dictionary object creator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PDFDict
}
\description{
PDF Dictionary object creator
}
\keyword{datasets}
