% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFStreamX.R
\docType{data}
\name{PDFLine}
\alias{PDFLine}
\title{PDF Line stream object creator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PDFLine
}
\description{
This is the R6 class representing the line stream object.
}
\details{
\itemize{
\item{x1,y1,x2,y2 pair of coordinates}
\item{fill,stroke colour specification}
\item{linewidth,linetype line specification}
\item{clip_rect clipping rectangle for this object.  Must be a vector of4 numbers giving x, y, width and height. default: NULL (no clipping).}
\item{... extra arguments ignored}
}
}
\keyword{datasets}
