% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFStreamX.R
\docType{data}
\name{PDFRect}
\alias{PDFRect}
\title{PDF Rect stream object creator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PDFRect
}
\description{
This is the R6 class representing the rect stream object.
}
\details{
\itemize{
\item{x,y coordinates}
\item{width,height rectangle width and height}
\item{fill,stroke colour specification}
\item{linewidth,linetype line specification}
\item{clip_rect clipping rectangle for this object.  Must be a vector of4 numbers giving x, y, width and height. default: NULL (no clipping).}
\item{transform transform specification. Default: NULL. Otherwise a character vector of transformation operations. See \code{tf} helper.}
\item{... extra arguments ignored}
}
}
\keyword{datasets}
