% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFStreamX.R
\docType{data}
\name{PDFText}
\alias{PDFText}
\title{PDF Text stream object creator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PDFText
}
\description{
This is the R6 class representing the text stream object.
}
\details{
\itemize{
\item{text text string for PDFText}
\item{x,y coordinates}
\item{... extra arguments specifying initial state e.g. 'fill'}
}
}
\keyword{datasets}
