% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics-state.R
\name{pgpar}
\alias{pgpar}
\title{Create graphical parameters for PDF objects}
\usage{
pgpar(
  col = "black",
  fill = "black",
  alpha = 1,
  lty,
  lwd,
  lineend,
  linejoin,
  linemitre,
  rule
)
}
\arguments{
\item{col, fill}{set graphics parameters for this object}

\item{alpha}{additional alpha applied to col, fill}

\item{lty, lwd, lineend, linejoin, linemitre}{line optins}

\item{rule}{fill rule. 'winding' (default) or 'evenodd'}
}
\value{
a graphics parameter object
}
\description{
This is similar to \code{grid::gpar()} except that values can only 
be scalars (i.e. length = 1)
}
\examples{
pgpar()
}
