% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif.R
\name{print.dict}
\alias{print.dict}
\title{Print a pif_dict. Can be nested}
\usage{
\method{print}{dict}(x, depth = 0, ...)
}
\arguments{
\item{x}{pif_dict}

\item{depth}{print depth. Default: 0.  Used to control indentation}

\item{...}{ignored}
}
\value{
None.
}
\description{
Print a pif_dict. Can be nested
}
