% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf-dict.R
\name{print.pdf_dict}
\alias{print.pdf_dict}
\title{Print a pdf_dict}
\usage{
\method{print}{pdf_dict}(x, ...)
}
\arguments{
\item{x}{pdf_dict}

\item{...}{ignored}
}
\value{
None.
}
\description{
Print a pdf_dict
}
