% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf-stream.R
\name{print.pdf_stream}
\alias{print.pdf_stream}
\title{Print \code{pdf_stream}}
\usage{
\method{print}{pdf_stream}(x, ...)
}
\arguments{
\item{x}{pdf_stream}

\item{...}{ignored}
}
\value{
None
}
\description{
Print \code{pdf_stream}
}
