% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFDocument.R
\name{process_fill_and_stroke}
\alias{process_fill_and_stroke}
\title{Process 'fill' and 'stroke' colours into snippets of the stream specification}
\usage{
process_fill_and_stroke(fill, stroke)
}
\arguments{
\item{fill, stroke}{4 element RGBA vectors}
}
\value{
named list with pre-rendered text streams for fill, stroke and
        the painting tpe to be used i.e.. 'f' = just fill, 'b' = fill and stroke,
        'S' =  just stroke?
}
\description{
Process 'fill' and 'stroke' colours into snippets of the stream specification
}
