% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-R6.R
\name{r6_to_funcs}
\alias{r6_to_funcs}
\title{Convert public methods in an R6 to wrapper functions}
\usage{
r6_to_funcs(r6class, envir = parent.env(), object_name = "r6object",
  include = NULL, exclude = c("clone", "initialize"), rename = NULL,
  silent = NULL)
}
\arguments{
\item{r6class}{the R6Class object}

\item{envir}{the environment in which to assign the wrapper functions}

\item{object_name}{the argument name as the object should appear in the functions}

\item{include, exclude}{names of methods to include/exclude}

\item{rename}{named vector to define name of function if it should be
different from the method name e.g. \code{c(func_name = "method_name")}}

\item{silent}{vector of names of R6 methods which should return invisible result.
default: NULL}
}
\description{
This function is inspired by the \code{R62S3} package.
}
