% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFDocument.R
\name{reset_clip_rect}
\alias{reset_clip_rect}
\title{Reset the clipping rectangle to encompass the entire page}
\arguments{
\item{pdfdoc}{PDF document. An R6 object of class \code{PDFDocument}. Created with \code{create_pdfdoc()} or \code{PDFDocument$new()}}
}
\value{
PDF document with clipping rectangle set added
}
\description{
Sets the global clipping rectangle to encompass the document
}
\section{R6 Usage}{
 $reset_clip_rect()
}

