% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFDocument.R
\name{sanitise_colour_to_rgba_vec}
\alias{sanitise_colour_to_rgba_vec}
\title{Sanitise colour to 4-element RGBA}
\usage{
sanitise_colour_to_rgba_vec(colour)
}
\arguments{
\item{colour}{specified as hex-colour (e.g. '#123456'),
standard colour name (e.g. 'red'), 3-element RGB vector (alpha will
be assumed to be 255), or 4-element RGB vector}
}
\value{
4 element vector representing RGBA with each element in range [0, 1]
}
\description{
When colour supplied as a vector, elements must in range [0, 255]
}
