% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFDocument.R
\name{set_clip_rect}
\alias{set_clip_rect}
\title{Set the clipping rectangle}
\arguments{
\item{pdfdoc}{PDF document. An R6 object of class \code{PDFDocument}. Created with \code{create_pdfdoc()} or \code{PDFDocument$new()}}

\item{x, y, width, height}{dimenensions of rectangle}
}
\value{
PDF document with clipping rectangle set added
}
\description{
Sets the global clipping rectangle for the document
}
\section{R6 Usage}{
 $set_clip_rect()
}

