% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-show.R
\name{show_pdf}
\alias{show_pdf}
\title{Show the PDF using the defined viewer}
\usage{
show_pdf(doc, viewer = getOption("viewer", utils::browseURL))
}
\arguments{
\item{doc}{pdf doc}

\item{viewer}{use the defined viewer. Default: utils::browseURL}
}
\value{
None
}
\description{
Show the PDF using the defined viewer
}
