% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-transformations.R
\docType{data}
\name{tf}
\alias{tf}
\title{Helper to create transformation strings}
\format{An object of class \code{list} of length 5.}
\usage{
tf
}
\description{
For details on PDF's coordinate transformations, see the standard
section 8.3.4 Transformation Matrices
}
\details{
\itemize{
\item{\code{translate(x, y)} - translate an object}
\item{\code{scale(sx, sy=sx)} - scale an object. If only 1 number is given, the object is scaled the same in both directions}
\item{\code{rotate(angle, x=0, y=0)} - rotate an object around the given point by the given angle (measured in degrees). By default rotate about the origin}
\item{\code{skew(A, B)} - skew an object with skew angles A and B (in degrees)}
\item{\code{custom(a, b, c, d, e, f)} - specify the 6 elements of a custom transformation matrix}
}
}
\keyword{datasets}
