% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{tf_translate}
\alias{tf_translate}
\title{Create a translation specification (for use as \code{tf} argument)}
\usage{
tf_translate(x, y)
}
\arguments{
\item{x, y}{translation}
}
\value{
translation specification
}
\description{
Create a translation specification (for use as \code{tf} argument)
}
\examples{
doc <- create_pdf() |>
   pdf_text(text = "hello", x = 0, y = 0, tf = tf_translate(x = 10, y = 10))
}
\seealso{
Other transform functions: 
\code{\link{pdf_rotate}()},
\code{\link{pdf_scale}()},
\code{\link{pdf_translate}()},
\code{\link{tf_rotate}()},
\code{\link{tf_scale}()}
}
\concept{transform functions}
