% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipping.R
\name{as.character.clip_rect}
\alias{as.character.clip_rect}
\alias{as.character.clip_polygon}
\title{Convert clipping spec into PDF string}
\usage{
\method{as.character}{clip_rect}(x, ...)

\method{as.character}{clip_polygon}(x, ...)
}
\arguments{
\item{x}{clip object}

\item{...}{ignored}
}
\description{
Convert clipping spec into PDF string
}
