% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{as.character.pdf_dict}
\alias{as.character.pdf_dict}
\title{Convert dict to character}
\usage{
\method{as.character}{pdf_dict}(x, depth = 0, ...)
}
\arguments{
\item{x}{pdf_dict}

\item{depth}{print depth. Default: 0.  Used to control indentation}

\item{...}{ignored}
}
\value{
None.
}
\description{
Convert dict to character
}
