% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipping.R
\name{clip_polygon}
\alias{clip_polygon}
\title{Define a clipping polygon}
\usage{
clip_polygon(xs, ys)
}
\arguments{
\item{xs, ys}{vertices}
}
\value{
clipping polygon specification
}
\description{
Define a clipping polygon
}
