% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipping.R
\name{clip_rect}
\alias{clip_rect}
\title{Define a clipping rectangle for use as a \code{clip} argument}
\usage{
clip_rect(x, y, width, height)
}
\arguments{
\item{x, y}{position}

\item{width, height}{size}
}
\value{
clipping rectangle specification
}
\description{
Define a clipping rectangle for use as a \code{clip} argument
}
\examples{
doc <- create_pdf() |>
   pdf_rect(0, 0, 100, 100, clip = clip_rect(50, 50, 200, 200))
}
