% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{create_pdf}
\alias{create_pdf}
\title{Create an empty shell for the PDF intermediate format}
\usage{
create_pdf(
  width = 400,
  height = 400,
  title = NULL,
  author = NULL,
  creator = "minipdf/R",
  creation_date = strftime(Sys.time(), format = "D:\%Y\%m\%d\%H\%M")
)
}
\arguments{
\item{width, height}{page size}

\item{title}{title}

\item{author}{author}

\item{creator}{creator}

\item{creation_date}{creation date}
}
\value{
List with attributes. List items are PDF objects.  Attributes
        are PDF settings
}
\description{
Create an empty shell for the PDF intermediate format
}
\examples{
create_pdf()

}
