% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minipdf-package.R
\docType{package}
\name{minipdf-package}
\alias{minipdf}
\alias{minipdf-package}
\title{minipdf: PDF Document Builder}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

PDF document builder.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coolbutuseless/minipdf}
}

}
\author{
\strong{Maintainer}: Mike Cheng \email{mikefc@coolbutuseless.com} [copyright holder]

}
\keyword{internal}
