% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf-stream.R
\name{pdf_circle}
\alias{pdf_circle}
\title{Create a polygon}
\usage{
pdf_circle(x, y, r, ..., gp = pgpar())
}
\arguments{
\item{x, y, r}{position and radius}

\item{...}{further arguments to be added to \code{gp}}

\item{gp}{A named list \code{gp} object created by \code{\link{pgpar}()}}
}
\value{
stream object
}
\description{
Create a polygon
}
