% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipping.R
\name{pdf_clip_polygon}
\alias{pdf_clip_polygon}
\title{Add a global clipping polygon}
\usage{
pdf_clip_polygon(doc, xs, ys, ..., gp = pgpar(), tf = NULL)
}
\arguments{
\item{doc}{pdf_doc}

\item{xs, ys}{vertices}

\item{...}{further arguments to be added to \code{gp}}

\item{gp}{A named list \code{gp} object created by \code{\link{pgpar}()}}

\item{tf}{either a single transform (\code{tf_translate()}, \code{tf_scale()},
\code{tf_rotate()}), or a list of these transforms.  Default: NULL,
no local transformation applied (global transformations still apply)}
}
\value{
pdf_doc
}
\description{
Clipping regions are cumulative, and these is no operation to expand the 
global clipping region.
use local clipping with the \code{clip} argument to individual objects.
}
