% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipping.R
\name{pdf_clip_polygon}
\alias{pdf_clip_polygon}
\title{Add a global clipping polygon}
\usage{
pdf_clip_polygon(doc, xs, ys, id = NULL, rule = "winding", tf = NULL)
}
\arguments{
\item{doc}{pdf_doc}

\item{xs, ys}{vertices}

\item{id}{A numeric vector used to searpate vertices into multiple polygons.
All vertices with the same id belong to the same polygon. Default: NULL
means that all vertices belong to a single polygon.}

\item{rule}{fill rule. 'winding' or 'evenodd'.  Default: 'winding'}

\item{tf}{either a single transform (\code{tf_translate()}, \code{tf_scale()},
\code{tf_rotate()}), or a list of these transforms.  Default: NULL,
no local transformation applied (global transformations still apply)}
}
\value{
pdf_doc
}
\description{
Clipping regions are cumulative, and these is no operation to expand the 
global clipping region.
use local clipping with the \code{clip} argument to individual objects.
}
