% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipping.R
\name{pdf_clip_rect}
\alias{pdf_clip_rect}
\title{Add a global clipping rectangle}
\usage{
pdf_clip_rect(doc, x, y, width, height, tf = NULL)
}
\arguments{
\item{doc}{pdf_doc}

\item{x, y}{position}

\item{width, height}{size}

\item{tf}{either a single transform (\code{tf_translate()}, \code{tf_scale()},
\code{tf_rotate()}), or a list of these transforms.  Default: NULL,
no local transformation applied (global transformations still apply)}
}
\value{
pdf_doc
}
\description{
Clipping regions are cumulative, and these is no operation to expand the 
global clipping region.
use local clipping with the \code{clip} argument to individual objects.
}
