% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipping.R
\name{pdf_clip_rect}
\alias{pdf_clip_rect}
\title{Add a rectangular clip path}
\usage{
pdf_clip_rect(doc, x, y, width, height, ..., gp = pgpar(), tf = NULL)
}
\arguments{
\item{doc}{pdf_doc}

\item{x, y}{position}

\item{width, height}{size}

\item{...}{further arguments to be added to \code{gp}}

\item{gp}{A named list \code{gp} object created by \code{\link{pgpar}()}}

\item{tf}{either a single transform \code{pdf_translate()}, \code{pdf_scale()},
\code{pdf_rotate()}, or a list of these transforms.  Default: NULL,
no transforms applied}
}
\value{
pdf_doc
}
\description{
Add a rectangular clip path
}
