% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf-stream.R
\name{pdf_image}
\alias{pdf_image}
\title{Add image}
\usage{
pdf_image(
  doc,
  im,
  x,
  y,
  scale = scale,
  interpolate = FALSE,
  ...,
  gp = pgpar(),
  tf = NULL,
  clip = NULL
)
}
\arguments{
\item{doc}{pdf_doc}

\item{im}{integer matrix [0, 255]}

\item{x, y}{position}

\item{scale}{scale for image}

\item{interpolate}{Should pixel values be interpolated? Default: FALSE}

\item{...}{further arguments to be added to \code{gp}}

\item{gp}{A named list \code{gp} object created by \code{\link{pgpar}()}}

\item{tf}{either a single transform (\code{tf_translate()}, \code{tf_scale()},
\code{tf_rotate()}), or a list of these transforms.  Default: NULL,
no local transformation applied (global transformations still apply)}

\item{clip}{either a single clip (\code{clip_rect()}, \code{clip_polygon()}),
or a list of these clips.  Default: NULL,
no local clipping applied (global clipping still applicable)}
}
\value{
pdf_doc
}
\description{
Add image
}
