% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf-stream.R
\name{pdf_image}
\alias{pdf_image}
\title{Add image}
\usage{
pdf_image(doc, im, x, y, scale = scale, ..., gp = pgpar())
}
\arguments{
\item{doc}{pdf_doc}

\item{im}{integer matrix [0, 255]}

\item{x, y}{position}

\item{scale}{scale for image}

\item{...}{further arguments to be added to \code{gp}}

\item{gp}{A named list \code{gp} object created by \code{\link{pgpar}()}}
}
\value{
pdf_doc
}
\description{
Add image
}
