% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf-stream.R
\name{pdf_line}
\alias{pdf_line}
\title{Create a line in pdf}
\usage{
pdf_line(x1, y1, x2, y2, ..., gp = list())
}
\arguments{
\item{x1, y1, x2, y2}{endpoints}

\item{...}{further arguments to be added to \code{gp}}

\item{gp}{A named list \code{gp} object created by \code{\link{pgpar}()}}
}
\value{
stream object
}
\description{
Create a line in pdf
}
