% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf-stream.R
\name{pdf_polygon}
\alias{pdf_polygon}
\title{Create a polygon}
\usage{
pdf_polygon(xs, ys, ..., gp = pgpar())
}
\arguments{
\item{xs, ys}{vertices}

\item{...}{further arguments to be added to \code{gp}}

\item{gp}{A named list \code{gp} object created by \code{\link{pgpar}()}}
}
\value{
stream object
}
\description{
Create a polygon
}
