% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf-stream.R
\name{pdf_polyline}
\alias{pdf_polyline}
\title{Create a polyline}
\usage{
pdf_polyline(doc, xs, ys, ..., gp = pgpar())
}
\arguments{
\item{doc}{pdf_doc}

\item{xs, ys}{vertices}

\item{...}{further arguments to be added to \code{gp}}

\item{gp}{A named list \code{gp} object created by \code{\link{pgpar}()}}
}
\value{
pdf_doc
}
\description{
Create a polyline
}
