% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf-stream.R
\name{pdf_rect}
\alias{pdf_rect}
\title{Create a rect}
\usage{
pdf_rect(x, y, width, height, ..., gp = list())
}
\arguments{
\item{x, y}{position}

\item{width, height}{size}

\item{...}{further arguments to be added to \code{gp}}

\item{gp}{A named list \code{gp} object created by \code{\link{pgpar}()}}
}
\value{
stream object
}
\description{
Create a rect
}
