% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf-stream.R
\name{pdf_rect}
\alias{pdf_rect}
\title{Create a rect}
\usage{
pdf_rect(x, y, width, height)
}
\arguments{
\item{x, y}{position}

\item{width, height}{size}
}
\value{
stream object
}
\description{
Create a rect
}
