% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{pdf_rotate}
\alias{pdf_rotate}
\title{Update global transformation matrix}
\usage{
pdf_rotate(doc, rads, x = 0, y = 0)
}
\arguments{
\item{doc}{pdf_doc}

\item{rads}{rotation angle in radians}

\item{x, y}{location to rotate around}
}
\value{
pdf_doc
}
\description{
Note: This can only be updated, and not reset.  If per-object transforms
      are desired use the \code{tf} argument on each object.
}
