% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{pdf_rotate}
\alias{pdf_rotate}
\title{build a transform object for rotation}
\usage{
pdf_rotate(rads, x = 0, y = 0)
}
\arguments{
\item{rads}{rotation angle in radians}

\item{x, y}{location to rotate around}
}
\value{
rotation object
}
\description{
build a transform object for rotation
}
