% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{pdf_rotate}
\alias{pdf_rotate}
\title{Modify global transformation matrix with new rotation}
\usage{
pdf_rotate(doc, rads, x = 0, y = 0)
}
\arguments{
\item{doc}{pdf_doc}

\item{rads}{rotation angle in radians}

\item{x, y}{location to rotate around}
}
\value{
pdf_doc
}
\description{
Global transformations are cumulative, and these is no operation to reset
the global transformation.
For local transformations use the \code{tf} argument for individual objects.
}
