% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{pdf_scale}
\alias{pdf_scale}
\title{Update global transformation matrix}
\usage{
pdf_scale(doc, x, y = x)
}
\arguments{
\item{doc}{pdf_doc}

\item{x, y}{scale amount in each direction. If 'y' value is not specified
it is made the same as the 'x' value}
}
\value{
pdf_doc
}
\description{
Note: This can only be updated, and not reset.  If per-object transforms
      are desired use the \code{tf} argument on each object.
}
