% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{pdf_scale}
\alias{pdf_scale}
\title{build a transform object for rotation}
\usage{
pdf_scale(x, y = x)
}
\arguments{
\item{x, y}{scale amount.}
}
\value{
scale object
}
\description{
build a transform object for rotation
}
