% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{pdf_scale}
\alias{pdf_scale}
\title{Modify global transformation matrix with new scaling}
\usage{
pdf_scale(doc, x, y = x)
}
\arguments{
\item{doc}{pdf_doc}

\item{x, y}{scale amount in each direction. If 'y' value is not specified
it is made the same as the 'x' value}
}
\value{
pdf_doc
}
\description{
Global transformations are cumulative, and these is no operation to reset
the global transformation.
For local transformations use the \code{tf} argument for individual objects.
}
