% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf-stream.R
\name{pdf_text}
\alias{pdf_text}
\title{Create text}
\usage{
pdf_text(text, x, y, fontsize = 12, mode = 0, ..., gp = pgpar())
}
\arguments{
\item{text}{text}

\item{x, y}{position}

\item{fontsize}{Default: 12}

\item{mode}{Default: 0
\itemize{
  \item{0 - Fill text. Normal. Default}
  \item{1 - Stroke text}
  \item{2 - Fill then stroke}
  \item{3 - NO fill or stroke. Invisible}
  \item{4 - Fill text and add to path for clipping}
  \item{5 - Stroke text and add to path for clipping}
  \item{6 - Fill, then stroke text and add to path for clipping}
  \item{7 - Add text to path for clipping}
}}

\item{...}{further arguments to be added to \code{gp}}

\item{gp}{A named list \code{gp} object created by \code{\link{pgpar}()}}
}
\value{
stream object
}
\description{
Create text
}
