% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf-stream.R
\name{pdf_text}
\alias{pdf_text}
\title{Add text to PDF}
\usage{
pdf_text(
  doc,
  text,
  x,
  y,
  fontsize = 12,
  mode = 0,
  ...,
  gp = pgpar(),
  tf = NULL,
  clip = NULL
)
}
\arguments{
\item{doc}{pdf_doc}

\item{text}{character string to render}

\item{x, y}{position}

\item{fontsize}{Default: 12}

\item{mode}{Default: 0
\itemize{
  \item{0 - Fill text. Normal. Default}
  \item{1 - Stroke text}
  \item{2 - Fill then stroke}
  \item{3 - NO fill or stroke. Invisible}
  \item{4 - Fill text and add to path for clipping}
  \item{5 - Stroke text and add to path for clipping}
  \item{6 - Fill, then stroke text and add to path for clipping}
  \item{7 - Add text to path for clipping}
}}

\item{...}{further arguments to be added to \code{gp}}

\item{gp}{A named list \code{gp} object created by \code{\link{pgpar}()}}

\item{tf}{either a single transform (\code{tf_translate()}, \code{tf_scale()},
\code{tf_rotate()}), or a list of these transforms.  Default: NULL,
no local transformation applied (global transformations still apply)}

\item{clip}{either a single clip (\code{clip_rect()}, \code{clip_polygon()}),
or a list of these clips.  Default: NULL,
no local clipping applied (global clipping still applicable)}
}
\value{
\code{pdf_doc}
}
\description{
Add text to PDF
}
\examples{
doc <- create_pdf() |>
   pdf_text("Hello", x = 20, y = 20, fontsize = 50)
}
