% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{pdf_translate}
\alias{pdf_translate}
\title{build a transform object for translation}
\usage{
pdf_translate(x, y)
}
\arguments{
\item{x, y}{translation}
}
\value{
translation object
}
\description{
build a transform object for translation
}
