% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{pdf_translate}
\alias{pdf_translate}
\title{Update global transformation matrix}
\usage{
pdf_translate(doc, x, y)
}
\arguments{
\item{doc}{pdf_doc}

\item{x, y}{translation}
}
\value{
pdf_doc
}
\description{
Note: This can only be updated, and not reset.  If per-object transforms
      are desired use the \code{tf} argument on each object.
}
