% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFStreamX.R
\docType{data}
\name{PDFCircle}
\alias{PDFCircle}
\title{PDF Circle stream object creator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PDFCircle
}
\description{
This is the R6 class representing the circle stream object.
}
\details{
\itemize{
\item{x,y coordinates}
\item{r radius of circle}
\item{... extra arguments specifying initial state e.g. 'fill'}
}
}
\keyword{datasets}
