% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFStreamX.R
\docType{data}
\name{PDFClipPolygon}
\alias{PDFClipPolygon}
\title{PDF ClipPolygon stream object creator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PDFClipPolygon
}
\description{
This is the R6 class representing the ClipPolygon stream object.
}
\details{
\itemize{
\item{xs,ys vectors of x and y coordinates}
\item{... extra arguments specifying initial state e.g. 'fill'}
}
}
\keyword{datasets}
