% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFStreamX.R
\docType{data}
\name{PDFCustom}
\alias{PDFCustom}
\title{PDF Custom stream object creator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PDFCustom
}
\description{
This is the R6 class representing the custom stream object.
}
\details{
\itemize{
\item{text text string for PDFText}
\item{new_graphics_state Should the object be drawn in its own local graphics state? default: TRUE}
\item{... extra arguments specifying initial state e.g. 'fill'}
}
}
\keyword{datasets}
