% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFDict.R
\name{PDFDict}
\alias{PDFDict}
\title{PDF Dictionary object creator}
\description{
PDF Dictionary object creator

PDF Dictionary object creator
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dict}}{TODO
Initialize a new PDF Dict}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PDFDict$new()}}
\item \href{#method-update}{\code{PDFDict$update()}}
\item \href{#method-as_character}{\code{PDFDict$as_character()}}
\item \href{#method-print}{\code{PDFDict$print()}}
\item \href{#method-as_object}{\code{PDFDict$as_object()}}
\item \href{#method-copy}{\code{PDFDict$copy()}}
\item \href{#method-clone}{\code{PDFDict$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFDict$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{named objects to add to dictionary
Update values in the dictionary.

Each named named argument to this
function is considered a name/value pair to include in the dictionary.
Unnamed arguments are silently dropped}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFDict$update(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{named objects to updated in dictionary
Character representation of this dict and all sub dictionaries it contains}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_character"></a>}}
\if{latex}{\out{\hypertarget{method-as_character}{}}}
\subsection{Method \code{as_character()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFDict$as_character(depth = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{depth}}{depth of dictionary currently being printed. default = 0}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
single character string
Print
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFDict$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{other arguments ignored
Convert this dict into a full PDF object representation

This object representation isfor inclusion in
a PDFDocument}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_object"></a>}}
\if{latex}{\out{\hypertarget{method-as_object}{}}}
\subsection{Method \code{as_object()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFDict$as_object(obj_idx)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj_idx}}{the object index to use within the document}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copy"></a>}}
\if{latex}{\out{\hypertarget{method-copy}{}}}
\subsection{Method \code{copy()}}{
Perform a proper deep copy of this object

Since PDFDict objects can contain other PDFDict objects in a list
need to do a deep copy to properly clone see private/deep_clone()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFDict$copy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFDict$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
