% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFDocument.R
\name{PDFDocument}
\alias{PDFDocument}
\title{PDF Document creator}
\description{
PDF Document creator

PDF Document creator
}
\details{
\itemize{
\item{\code{fontname} - default: Helvetica. Choose one of the standard PDF
fonts: Times-Roman, Times-Bold,
Times-Italic, Times-BoldItalic, Helvetica, Helvetica-Bold, Helvetica-Oblique,
Helvetica-BoldOblique, Courier, Courier-Bold, Courier-Oblique, Courier-BoldOblique,
Symbol, ZapfDingbats}
\item{\code{width.height} document dimensions. Default 400x400}
\item{\code{version} Default: 1.2.  Doesn't do anything other than set the string at the
       beginning of the PDF document.}
\item{\code{...} PDF objects}
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{version}}{PDF spec version}

\item{\code{width, height}}{document width and height}

\item{\code{user_objects}}{User elements added to the document}

\item{\code{setup_objects}}{standard PDF header objects not directly defined
       by the user
Initialize a new PDF document

Configure the setup objects and initialise the user_objects
The setup objects are those PDF objects needed to initialise the page
at the most basic level i.e.
\itemize{
\item{set up a catalog}
\item{set up a list of pages}
\item{set up font information.}
\item{set up the list of page contents (which will get updated with each new object)}
\item{keep track of GraphicStates used by various objects}
}}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{width, height}}{document width and height}

\item{\code{all_objects}}{List containing all objects}

\item{\code{objs}}{All strings for all objects}

\item{\code{header}}{The header for this document}

\item{\code{body}}{The body is just the collection of all objects}

\item{\code{xref}}{Update a xref table based upon the current objects

and the offsets to the start of each object
Each xref entry must be exactly 20bytes (include CRLF).  Since I'm only
using CR, I need to add an extra space at the end of each xref entry}

\item{\code{offsets}}{For the XREF table, need to know the offset to each object from start of document}

\item{\code{trailer}}{The trailer indicates the start of the XREF table and total length
of all objects it represents (including the dummy/root object)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PDFDocument-new}{\code{PDFDocument$new()}}
\item \href{#method-PDFDocument-append}{\code{PDFDocument$append()}}
\item \href{#method-PDFDocument-replace}{\code{PDFDocument$replace()}}
\item \href{#method-PDFDocument-delete}{\code{PDFDocument$delete()}}
\item \href{#method-PDFDocument-as_character}{\code{PDFDocument$as_character()}}
\item \href{#method-PDFDocument-print}{\code{PDFDocument$print()}}
\item \href{#method-PDFDocument-save}{\code{PDFDocument$save()}}
\item \href{#method-PDFDocument-copy}{\code{PDFDocument$copy()}}
\item \href{#method-PDFDocument-show}{\code{PDFDocument$show()}}
\item \href{#method-PDFDocument-text}{\code{PDFDocument$text()}}
\item \href{#method-PDFDocument-rect}{\code{PDFDocument$rect()}}
\item \href{#method-PDFDocument-line}{\code{PDFDocument$line()}}
\item \href{#method-PDFDocument-circle}{\code{PDFDocument$circle()}}
\item \href{#method-PDFDocument-polygon}{\code{PDFDocument$polygon()}}
\item \href{#method-PDFDocument-polyline}{\code{PDFDocument$polyline()}}
\item \href{#method-PDFDocument-custom}{\code{PDFDocument$custom()}}
\item \href{#method-PDFDocument-dict}{\code{PDFDocument$dict()}}
\item \href{#method-PDFDocument-clone}{\code{PDFDocument$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFDocument-new"></a>}}
\if{latex}{\out{\hypertarget{method-PDFDocument-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFDocument$new(
  ...,
  width = 400,
  height = 400,
  fontname = "Helvetica",
  version = 1.2
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{user objects to add during initialisation}

\item{\code{width, height}}{dimensions of PDF document}

\item{\code{fontname}}{default font for document (Default: 'Helvetica')}

\item{\code{version}}{PDF specifcation version. Default: 1.2
Append user-defined objects to document}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFDocument-append"></a>}}
\if{latex}{\out{\hypertarget{method-PDFDocument-append}{}}}
\subsection{Method \code{append()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFDocument$append(..., position = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{objects}

\item{\code{position}}{where should the objects be inserted in the user object
       list?  Default: NULL means to append at end of object list.
       A value of "1" means to insert at start of object list.
Replace an object at the given position}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFDocument-replace"></a>}}
\if{latex}{\out{\hypertarget{method-PDFDocument-replace}{}}}
\subsection{Method \code{replace()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFDocument$replace(pdf_object, position)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pdf_object}}{User defined PDF object e.g. TODO}

\item{\code{position}}{numeric index in object list
Delete the object at the given position}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFDocument-delete"></a>}}
\if{latex}{\out{\hypertarget{method-PDFDocument-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFDocument$delete(position)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{position}}{numeric index in object list
Character reprsetntatino of PDF document as a single string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFDocument-as_character"></a>}}
\if{latex}{\out{\hypertarget{method-PDFDocument-as_character}{}}}
\subsection{Method \code{as_character()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFDocument$as_character(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored
Print}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFDocument-print"></a>}}
\if{latex}{\out{\hypertarget{method-PDFDocument-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFDocument$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored
Save to PDF file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFDocument-save"></a>}}
\if{latex}{\out{\hypertarget{method-PDFDocument-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFDocument$save(filename, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{output filename}

\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFDocument-copy"></a>}}
\if{latex}{\out{\hypertarget{method-PDFDocument-copy}{}}}
\subsection{Method \code{copy()}}{
Deep copy of this document
Show the PDF in the defined R viewer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFDocument$copy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFDocument-show"></a>}}
\if{latex}{\out{\hypertarget{method-PDFDocument-show}{}}}
\subsection{Method \code{show()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFDocument$show(viewer = getOption("viewer", utils::browseURL))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{viewer}}{which viewer to use}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFDocument-text"></a>}}
\if{latex}{\out{\hypertarget{method-PDFDocument-text}{}}}
\subsection{Method \code{text()}}{
Initialize a stream object representing a text
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFDocument$text(text, x, y, fontsize, text_mode, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{text}

\item{\code{x, y}}{location}

\item{\code{fontsize, text_mode}}{text settings [optional]}

\item{\code{...}}{initial named attributes of this object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFDocument-rect"></a>}}
\if{latex}{\out{\hypertarget{method-PDFDocument-rect}{}}}
\subsection{Method \code{rect()}}{
Initialize a stream object representing a rectangle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFDocument$rect(x, y, width, height, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x, y, width, height}}{specificaiton of rectangle extents}

\item{\code{...}}{initial named attributes of this object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFDocument-line"></a>}}
\if{latex}{\out{\hypertarget{method-PDFDocument-line}{}}}
\subsection{Method \code{line()}}{
Initialize a stream object representing a line
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFDocument$line(x1, y1, x2, y2, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x1, y1, x2, y2}}{line start/end coordinates}

\item{\code{...}}{initial named attributes of this object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFDocument-circle"></a>}}
\if{latex}{\out{\hypertarget{method-PDFDocument-circle}{}}}
\subsection{Method \code{circle()}}{
Initialize a stream object representing a circle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFDocument$circle(x, y, r, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x, y}}{centre of circle}

\item{\code{r}}{radius}

\item{\code{...}}{initial named attributes of this object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFDocument-polygon"></a>}}
\if{latex}{\out{\hypertarget{method-PDFDocument-polygon}{}}}
\subsection{Method \code{polygon()}}{
Initialize a stream object representing a polygon
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFDocument$polygon(xs, ys, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xs, ys}}{coordinates of polygon vertices}

\item{\code{...}}{initial named attributes of this object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFDocument-polyline"></a>}}
\if{latex}{\out{\hypertarget{method-PDFDocument-polyline}{}}}
\subsection{Method \code{polyline()}}{
Initialize a stream object representing a polyline
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFDocument$polyline(xs, ys, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xs, ys}}{numeric vectors of x, ycoordinates along polyline}

\item{\code{...}}{initial named attributes of this object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFDocument-custom"></a>}}
\if{latex}{\out{\hypertarget{method-PDFDocument-custom}{}}}
\subsection{Method \code{custom()}}{
Initialize a stream object representing a custom object with text
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFDocument$custom(text, new_graphics_state = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{text string}

\item{\code{new_graphics_state}}{Should the object be drawn in its own local
graphics state? default: TRUE}

\item{\code{...}}{initial named attributes of this object
Initialize a new PDF Dict}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFDocument-dict"></a>}}
\if{latex}{\out{\hypertarget{method-PDFDocument-dict}{}}}
\subsection{Method \code{dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFDocument$dict(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{named objects to add to dictionary}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PDFDocument-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PDFDocument-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFDocument$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
