% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFDocument.R
\docType{data}
\name{PDFDocument}
\alias{PDFDocument}
\title{PDF Document creator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PDFDocument
}
\description{
\itemize{
\item{\code{fontname} - default: Helvetica. Choose one of the standard PDF
fonts: Times-Roman, Times-Bold,
Times-Italic, Times-BoldItalic, Helvetica, Helvetica-Bold, Helvetica-Oblique,
Helvetica-BoldOblique, Courier, Courier-Bold, Courier-Oblique, Courier-BoldOblique,
Symbol, ZapfDingbats}
\item{\code{width.height} document dimensions. Default 400x400}
\item{\code{version} Default: 1.2.  Doesn't do anything other than set the string at the
       beginning of the PDF document.}
\item{\code{...} PDF objects}
}
}
\keyword{datasets}
