% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFStreamX.R
\docType{data}
\name{PDFLine}
\alias{PDFLine}
\title{PDF Line stream object creator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PDFLine
}
\description{
This is the R6 class representing the line stream object.
}
\details{
\itemize{
\item{x1,y1,x2,y2 pair of coordinates}
\item{... extra arguments specifying initial state e.g. 'fill'}
}
}
\keyword{datasets}
