% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFStreamX.R
\docType{data}
\name{PDFPolygon}
\alias{PDFPolygon}
\title{PDF Polygon stream object creator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PDFPolygon
}
\description{
This is the R6 class representing the polygon stream object.
}
\details{
\itemize{
\item{xs,ys vectors of x and y coordinates}
\item{fill,stroke colour specification}
\item{linewidth,linetype line specification}
\item{clip_rect clipping rectangle for this object.  Must be a vector of 4 numbers giving x, y, width and height. default: NULL (no clipping).}
\item{... extra arguments ignored}
}
}
\keyword{datasets}
