% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFStreamX.R
\docType{data}
\name{PDFPolyline}
\alias{PDFPolyline}
\title{PDF Polyline stream object creator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PDFPolyline
}
\description{
This is the R6 class representing the polyline stream object.
}
\details{
\itemize{
\item{x,y coordinates}
\item{... extra arguments specifying initial state e.g. 'fill'}
}
}
\keyword{datasets}
