% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFStream.R
\docType{data}
\name{PDFStream}
\alias{PDFStream}
\title{PDF Stream Object Base Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PDFStream
}
\description{
PDF Stream Object Base Class
}
\keyword{datasets}
