% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFStream.R
\name{PDFStream}
\alias{PDFStream}
\title{PDF Stream Object Base Class}
\description{
PDF Stream Object Base Class

PDF Stream Object Base Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{attrib}}{TODO}

\item{\code{transform}}{TODO}

\item{\code{parent_attrib}}{TODO}

\item{\code{parent_transform}}{TODO
Initialize a stream object}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{alphas}}{all alpha channels}

\item{\code{gs_name}}{current GS name TODO: insert PDF spec reference}

\item{\code{gs_spec}}{current gs_spec}

\item{\code{clipping_spec}}{current clipping spec
Use clip_polygon if specified.
Use clip_rect if specified
else NULL}

\item{\code{paint_spec}}{path painting style based upon which of fill/stroke are NULL
One of 'n', 's', 'f', 'b'}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PDFStream$new()}}
\item \href{#method-update}{\code{PDFStream$update()}}
\item \href{#method-copy}{\code{PDFStream$copy()}}
\item \href{#method-as_object}{\code{PDFStream$as_object()}}
\item \href{#method-print}{\code{PDFStream$print()}}
\item \href{#method-rotate}{\code{PDFStream$rotate()}}
\item \href{#method-translate}{\code{PDFStream$translate()}}
\item \href{#method-scale}{\code{PDFStream$scale()}}
\item \href{#method-fontsize}{\code{PDFStream$fontsize()}}
\item \href{#method-text_mode}{\code{PDFStream$text_mode()}}
\item \href{#method-fill}{\code{PDFStream$fill()}}
\item \href{#method-stroke}{\code{PDFStream$stroke()}}
\item \href{#method-linewidth}{\code{PDFStream$linewidth()}}
\item \href{#method-linetype}{\code{PDFStream$linetype()}}
\item \href{#method-clip_rect}{\code{PDFStream$clip_rect()}}
\item \href{#method-clip_polygon}{\code{PDFStream$clip_polygon()}}
\item \href{#method-as_character}{\code{PDFStream$as_character()}}
\item \href{#method-get_transform_spec}{\code{PDFStream$get_transform_spec()}}
\item \href{#method-get_attrib}{\code{PDFStream$get_attrib()}}
\item \href{#method-clone}{\code{PDFStream$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFStream$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{stream attributes
Update the attributes of an object.

e.g. update the x coordinate:
\code{obj$update(x = 12)}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFStream$update(
  ...,
  fontsize,
  text_mode,
  fill,
  stroke,
  linewidth,
  linetype,
  clip_rect
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{named attributes}

\item{\code{fontsize, text_mode, fill, stroke, linewidth, linetype, clip_rect}}{common
drawing attributes}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copy"></a>}}
\if{latex}{\out{\hypertarget{method-copy}{}}}
\subsection{Method \code{copy()}}{
Clone the object
Convert to full PDF object representation include stream/endstream
and length dict.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFStream$copy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_object"></a>}}
\if{latex}{\out{\hypertarget{method-as_object}{}}}
\subsection{Method \code{as_object()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFStream$as_object(obj_idx)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obj_idx}}{the index to use for this object in the PDF document}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Cat just the inner stream part of the object when print()ed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFStream$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{arguments passed to \code{PDFStream$as_character(...)}
Transformations: rotate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rotate"></a>}}
\if{latex}{\out{\hypertarget{method-rotate}{}}}
\subsection{Method \code{rotate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFStream$rotate(degrees, x = 0, y = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{degrees}}{angle of rotation}

\item{\code{x, y}}{rotate around this point. default (0,1)
Transformations: translate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-translate"></a>}}
\if{latex}{\out{\hypertarget{method-translate}{}}}
\subsection{Method \code{translate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFStream$translate(x, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x, y}}{translation distance
Transformations: scale}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scale"></a>}}
\if{latex}{\out{\hypertarget{method-scale}{}}}
\subsection{Method \code{scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFStream$scale(x, y = x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x, y}}{scale factors in x and y directions. If only \code{x} scale
       is given, then this value will also be used for the \code{y} scale factor
Convenience method for changing core attribute: fontsize}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fontsize"></a>}}
\if{latex}{\out{\hypertarget{method-fontsize}{}}}
\subsection{Method \code{fontsize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFStream$fontsize(fontsize = 12)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fontsize}}{fontsize default: 12
Convenience method for changing core attribute: text_mode}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-text_mode"></a>}}
\if{latex}{\out{\hypertarget{method-text_mode}{}}}
\subsection{Method \code{text_mode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFStream$text_mode(text_mode = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text_mode}}{text mode. default 0.  TODO: insert reference to PDF spec
Convenience method for changing core attribute: fill colour}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fill"></a>}}
\if{latex}{\out{\hypertarget{method-fill}{}}}
\subsection{Method \code{fill()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFStream$fill(fill = "#000000")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fill}}{fill colour. default '#000000'
Convenience method for changing core attribute: stroke colour}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stroke"></a>}}
\if{latex}{\out{\hypertarget{method-stroke}{}}}
\subsection{Method \code{stroke()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFStream$stroke(stroke = "#000000")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stroke}}{stroke colour. default: '#000000'
Convenience method for changing core attribute: linewidth}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-linewidth"></a>}}
\if{latex}{\out{\hypertarget{method-linewidth}{}}}
\subsection{Method \code{linewidth()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFStream$linewidth(linewidth = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{linewidth}}{linewidth. default: 1
Convenience method for changing core attribute: fontsize}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-linetype"></a>}}
\if{latex}{\out{\hypertarget{method-linetype}{}}}
\subsection{Method \code{linetype()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFStream$linetype(linetype = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{linetype}}{linetype. default: 0. TODO: Insert reference to PDF spec
Convenience method for changing core attribute: clipping rect}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clip_rect"></a>}}
\if{latex}{\out{\hypertarget{method-clip_rect}{}}}
\subsection{Method \code{clip_rect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFStream$clip_rect(x, y, width, height)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x, y, width, height}}{dimensions of current clipping rect
Convenience method for changing core attribute: clip polygon}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clip_polygon"></a>}}
\if{latex}{\out{\hypertarget{method-clip_polygon}{}}}
\subsection{Method \code{clip_polygon()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFStream$clip_polygon(xs, ys)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xs, ys}}{coords of clipping polygon
convert stream object to character string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_character"></a>}}
\if{latex}{\out{\hypertarget{method-as_character}{}}}
\subsection{Method \code{as_character()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFStream$as_character(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_transform_spec"></a>}}
\if{latex}{\out{\hypertarget{method-get_transform_spec}{}}}
\subsection{Method \code{get_transform_spec()}}{
Full transform of all parent objects and this object's transform
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFStream$get_transform_spec()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_attrib"></a>}}
\if{latex}{\out{\hypertarget{method-get_attrib}{}}}
\subsection{Method \code{get_attrib()}}{
Normalised full set of attributes including parent attributes if available
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFStream$get_attrib()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PDFStream$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
