% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFStreamX.R
\docType{data}
\name{PDFText}
\alias{PDFText}
\title{PDF Text stream object creator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PDFText
}
\description{
This is the R6 class representing the text stream object.
}
\details{
\itemize{
\item{text text string for PDFText}
\item{x,y coordinates}
\item{fontsize font size}
\item{text_mode Integer. 0 = fill, 1 = stroke, 2 = fill and stroke. Default: 0}
\item{fill,stroke colour specification}
\item{linewidth,linetype line specification}
\item{clip_rect clipping rectangle for this object.  Must be a vector of4 numbers giving x, y, width and height. default: NULL (no clipping).}
\item{... extra arguments ignored}
}
}
\keyword{datasets}
