% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics-state.R
\name{pgpar}
\alias{pgpar}
\title{Create graphical parameters}
\usage{
pgpar(col = "black", fill = NA, alpha = 1, rule = "winding")
}
\arguments{
\item{col, fill}{set graphics parameters for this object}

\item{alpha, rule}{options}
}
\value{
a graphics parameter object
}
\description{
Create graphical parameters
}
\examples{
pgpar()
}
