% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{print.pdf_dict}
\alias{print.pdf_dict}
\title{Print a pdf Can be nested}
\usage{
\method{print}{pdf_dict}(x, depth = 0, ...)
}
\arguments{
\item{x}{pdf_dict}

\item{depth}{print depth. Default: 0.  Used to control indentation}

\item{...}{ignored}
}
\value{
None.
}
\description{
Print a pdf Can be nested
}
