% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{print.pdf_doc}
\alias{print.pdf_doc}
\title{Print a 'pdf' object}
\usage{
\method{print}{pdf_doc}(x, ...)
}
\arguments{
\item{x}{pdf object}

\item{...}{ignored}
}
\value{
None
}
\description{
Print a 'pdf' object
}
