% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{tf_rotate}
\alias{tf_rotate}
\title{Create a transform specification for rotation}
\usage{
tf_rotate(rads, x = 0, y = 0)
}
\arguments{
\item{rads}{rotation angle in radians}

\item{x, y}{location to rotate around}
}
\value{
rotation specification
}
\description{
Create a transform specification for rotation
}
