% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{tf_scale}
\alias{tf_scale}
\title{Create a transform specification for scaling}
\usage{
tf_scale(x, y = x)
}
\arguments{
\item{x, y}{scale amount in each direction. If 'y' value is not specified
it is made the same as the 'x' value}
}
\value{
scale specification
}
\description{
Create a transform specification for scaling
}
