% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{tf_scale}
\alias{tf_scale}
\title{Create a scaling specification (for use as \code{tf} argument)}
\usage{
tf_scale(x, y = x)
}
\arguments{
\item{x, y}{scale amount in each direction. If 'y' value is not specified
it is made the same as the 'x' value}
}
\value{
scale transform specification
}
\description{
Create a scaling specification (for use as \code{tf} argument)
}
\examples{
doc <- create_pdf() |>
   pdf_text(text = "hello", x = 0, y = 0, tf = tf_scale(x = 10))
}
