% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{tf_translate}
\alias{tf_translate}
\title{Create a transform specification for translation}
\usage{
tf_translate(x, y)
}
\arguments{
\item{x, y}{translation}
}
\value{
translation specification
}
\description{
Create a transform specification for translation
}
