% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDFDocument.R
\name{write_pdf}
\alias{write_pdf}
\title{Write PDF document to file}
\arguments{
\item{pdfdoc}{PDF document. An R6 object of class \code{PDFDocument}. Created with \code{create_pdfdoc()} or \code{PDFDocument$new()}}

\item{filename}{name of output file}
}
\description{
Write the PDF document to the given filename.
}
\section{R6 Usage}{
 $write_pdf()
}

