% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif.R
\name{write_pdf}
\alias{write_pdf}
\title{Write from the intermediate format to an actual PDF file}
\usage{
write_pdf(pif, filename)
}
\arguments{
\item{pif}{pif object as created by \code{\link{create_pif}()}}

\item{filename}{output filename}
}
\value{
None
}
\description{
Write from the intermediate format to an actual PDF file
}
