% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{write_pdf}
\alias{write_pdf}
\title{Write pdf to file or string}
\usage{
write_pdf(doc, filename = NULL)
}
\arguments{
\item{doc}{pdf_doc}

\item{filename}{Output filename. Default: NULL  no output to file but return
a string representation of the PDF}
}
\value{
string or None
}
\description{
Write pdf to file or string
}
\examples{
create_pdf() |>
   pdf_circle(200, 200, 50, lwd = 5, fill = 'hotpink') |>
   write_pdf() |>
   cat()
}
